# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoLang - Gin - RESTful - MongoDB - Redis",
        "title": "GoodBlast API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Muhammet Kerem Ozturk",
            "url": "https://github.com/kerem-ozt",
            "email": "mkeremozt@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
...
        "/user/whoami": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets current users info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Who Am I",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Switch page by 'page'",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/user/getall": {
            "get": {
                "description": "gets all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get All Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Switch page by 'page'",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Switch limit by 'limit'",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "users": [
                                        {
                                            "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                            "email": "test@gmail.com",
                                            "name": "test",
                                            "score": 0,
                                            "country": "Turkey",
                                            "createdAt": "2020-10-30T00:00:00Z",
                                            "updatedAt": "2020-10-30T00:00:00Z"
                                        }
                                    ],
                                    "prev": false,
                                    "next": false
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
        "/user/getbyid": {
            "get": {
                "description": "gets user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get User By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "user": {
                                        "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                        "email": "test@gmail.com",
                                        "name": "test",
                                        "score": 0,
                                        "country": "Turkey",
                                        "createdAt": "2020-10-30T00:00:00Z",
                                        "updatedAt": "2020-10-30T00:00:00Z"
                                    }
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
        "/user/delete": {
            "delete": {
                "description": "deletes user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete User By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/user/entertournament": {
            "post": {
                "description": "enter tournament",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Enter Tournament",
                "parameters": [
                    {
                        "description": "Enter Tournament Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EnterTournamentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Bearer-Token",
            "in": "header"
        }
    }
}